+++ release_notes.md
@@ -3,4 +3,4 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-- Add additive platform notification for sync triggers under `FUNCTIONS_SYNC_TRIGGERS_MODE=Platform` (#11813)
+- Add additive platform notification for sync triggers under `FUNCTIONS_NOTIFY_PLATFORM_ON_SYNC=true` (#11813)